Deploy ARA API server on bridge.sf.io
This retrieves ARA from source and puts it in the bridge.sf.io Ansible
role path. The ara_api [1] role is then eventually applied to install
ara in a virtualenv where it will be setup to run as a gunicorn service.

The API server will only listen to localhost on purpose for the time
being.

[1]: https://ara.readthedocs.io/en/latest/ansible-role-ara-api.html

Change-Id: If82e96c95ee1b4e96a398eac6d0b6ddcf3a065a9