Enable systemd service monitoring on the node-exporter-rpm role
In some cases, we will want to monitor the status of systemd services
in our nodes. Node Exporter provides such functionality, but it is not
enabled by default.

This commit adds a variable to the role, named
prometheus_monitored_services. When specified, it will modify file
/etc/sysconfig/node_exporter to enable systemd monitoring, and whitelist
those servers that are specified in the list.

Change-Id: I646fe7d455ee73e6e72c7e02c949823dd082dcbf