Add monitoring of AFS mirror sync status
We want to monitor the status of the openstack-infra AFS sync status
for the CentOS repo, as a sync failure might affect some of our CI
jobs.

This commit adds a role and alerting rule to monitor the sync status.
Currently, only the status of the CentOS sync is checked, but we can
add further checks in the future.

Change-Id: I44c3bcd924910df0e018e7e8822012f77454b967