Add DLRN admin users to dlrn-server role
Even though we are planning to do all our admin operations using
Git, for the DLRN server use case we will need to ocassionally
have some users connect to the VMs. This will be required to recheck
failed builds or troubleshoot failed synchronizations.

This commit adds some users who are authorized to connect to the
DLRN server and run local operations. These users should take care
not to do and admin operation that should be handled via Git, since
they would be overruled anyway. For this, it makes use of the users.yml
playbook from the rdo-infra-playbooks/base role.

Change-Id: I4ab8bc27507dfb8337344fc9d3843d804464e027