Run openshift-ansible's deploy_cluster playbook on registry node
This prepares the openshift-ansible roles and playbooks in the right
path, sets up groups expected by the openshift-ansible playbooks and
finally runs the deploy_cluster playbook after running pre-requisite
playbooks.

We also needed to move variables from group_vars to the inventory
because the openshift-ansible playbooks attempt to retrieve them
relative to the location of the playbooks and that's not where our
group_vars are.

It also turns out that openshift-ansible seeks SSL certificates on the
Ansible control node rather than the OpenShift node.
We've downloaded them to the bridge server to unblock this for now.

Change-Id: Ife57fbb7e54db2407a8aa3da62a422e93875e69b
Depends-On: https://review.rdoproject.org/r/24422
Depends-On: https://softwarefactory-project.io/r/#/c/17678/
Depends-On: https://softwarefactory-project.io/r/#/c/17736/
Depends-On: https://review.rdoproject.org/r/#/c/25947/