Fix reported dates for AFS, backup and DLRN alerts
The $value received by the alerts is whatever expression we use in
the expr section in the alert configuration. We need to change the
expression so that $value represents the actual last time for the
event, instead of the difference.

Also, we do not want to warn when backups are more than 1 day old. We
run backups every two days, so we should treigger an alert if it is
3 or more days old.

Change-Id: I915191c2d8b0192ac2efb3413e14fefaba402048