infra: refactors vars configuration using higher order Instance type
The main goal of this change is to remove from the `playbooks/vars` files:

- Ansible connection
- groups
- host vars
- Extenal servers

This results in avoiding full tenant reconfiguration when one of the above
item is changed.

To do that effectively, this change introduces a new Instance type
to decouple server, connection and group attributes as well as to
link a server with a volume.

The secondary goal of this change is to provide a good foundation and enables
efficient re-use of the `conf` directory at the cost of a major refactor.
To effectively review this change, we only need to verify the YAML file
is not unexpectedly altered.

Change-Id: Ica7ddda4aa1f707a8b13a24604844a77f042f09e