Adding third volume to logserver
We would like to keep logs for 15 days. In current situation,
we use over 1.8 TB in 14 days, but in the mean time
ci.centos.org infra has been added to the trusted host and it also
send the logs to the host.
It will be more secure to have 3 volumes in this server before
we reduce amount of logs sending to the host.

TG-3401

Change-Id: I5567c8edf8aab1c782921a38ca9372aa8139274f