Added new systemd services to monitoring by prometheus
In previous patch set [1], systemd services has been added
correctly as a whitelist services for node_exporter service,
but the handler doesn't have enough permissions to restart
node_exporter on all servers. After patching that,
handler was not triggered, so node_exporter needs to be restarted
on all servers.
This commit is adding new services, so node_exporter will be
restarted on all hosts.

[1] https://softwarefactory-project.io/r/#/c/18023/

Change-Id: I17f7cdf8343539c1b60dc07f546de77d259963da