Use puppet-dlrn from Zuul in sf-infra-molecule job
We want to integrate puppet-dlrn into sf-infra jobs. This initial step
allows us to use the puppet-dlrn repo created by Zuul in the molecule
job.

With this, we can use depends-on on puppet-dlrn patches, and also add
an sf-infra-molecule job as 3rd party CI to puppet-dlrn jobs (this should
be a follow-up patch).

Change-Id: I09ee9e55132e2e736a1583558d3ea70e59b345cd