Deploy an ARA API server container
This leverages the same approach as other services deployed in
containers such as prometheus for standing up a container image.

It results in an ARA API server running from a container in podman.
It is configured to use the default sqlite database backend for now
but can be made to connect to a mysql or postgresql server later.

The API server is exposed by nginx where SSL termination is also done.

This server will make it possible to send data from bridge.sf.io as
well as jobs.

Change-Id: I9c6b33928fda19806b3d2089a0953c3e49f6e3a7