Disable DLRN periodic runs for centos7 pike and ocata builders
Pike and Ocata releases have been in Extended Maintenance for long time
and there is no activity in those branches now, so I think we can remove
RDO Trunk repos.

In this patch, I'm disabling cron jobs for both DLRN executions and
repos purge. For removal, it will be properly announced in RDO and
removed later.

Change-Id: Iac6db975900beba79fced9d09b8bc014cf79d67e