ansible: replace module_default statement
The module_default statement doesn't work well with ansible prior to 2.8 so it's
difficult to reuse playbooks and roles from sf-infra on centos.

This change define auth parameters for os_* roles and modify playbooks to use a
dict which contain cloud and project name.

Also, it's possible to have a mix of clouds if needed (useful with app
credential for an app account is only usable in a tenant)

Change-Id: Ie43d2ce1d7fbcbb0377837ef60b0dd559bdc9fc9