Changing new workflow of replacing SSL certs in Openshift
We need to change the way of deploying SSL certificates in
Openshift to avoid self-signed certs.
This commit is changing playbook to use official way
of deploying SSL certificates. [1]
NOTE: The cluster certificates will be still self signed in,
but all configuration will be recreated, so adding
insecure-skip-tls-verify parameter will be not required.

[1] https://docs.openshift.com/container-platform/3.11/install_config/redeploying_certificates.html

Depends-On: https://review.rdoproject.org/r/#/c/28459/

Change-Id: Ifa7b8537fbfe937ded48909f9907fd131de6fd28