Add firewalld rules for Prometheus host
After a reboot, we see firewalld is enabled on the Prometheus VM, so
we need to add the required firewall rules to allow communication.

This also requires a fix to the firewalld role, since we now need to
specify a source host for some rules. In that case, we will use a rich
rule.

Change-Id: I52f4381c113b16a7853c0d4e3fab31fccd068596