Setup Prometheus service with defined image sha256 digest
We would like to run service in the latest version. Normally,
when image has been downloaded with the latest tag, it will be
not updated when the new one is available in the repository.
Thats why setting sha256 digest will be helpful to provide new
service version.
Also with that commit, if some pod_params are modified for
container service, ansible will recreate the container.

Change-Id: I3c6d3988d8576d23fc911a0b7c4ab965090ed5d7