Sync the src directory when running configure-registry-host
Previously, we did not make sure the src directory was synced from
the Zuul executor to the bridge when running configure-registry-host.
It used to work, simply because we were running other jobs before
that took care of that. But now, when running it in the thirdparty-post
pipeline, this is not always the case.

We fix this by making sure the ~/src directory is synced from the
Zuul executor, just like it is done in the sf-infra-configure-hosts
job.

Change-Id: I4cfb7e811042fdbe30d7a474782563952c53dee1