Ensure puppet < 6.0 is installed on DLRN hosts
We just got puppet 6.14.0 in the EPEL8 repository. This version is
breaking puppet-dlrn in various ways, starting by the lack of
"puppet module build" support, which is taken to the PDK... But this
package does not seem to be in EPEL.

Let's ensure we are installing a 5.x version until we get puppet-dlrn
to support Puppet 6, or migrate to Ansible.

Change-Id: I9c3849b3eae91fba135064dd9205cf93bdc78e10