Add afs-mirror-new role and forked roles from system-config
We want to update the AFS mirrors we are managing, to make sure
they mimic the setup used by the OpenDev infra. Since the previous
puppet-based setup is now deprecated, we are moving to the
Ansible-based roles.

Since the changes in to support CentOS in the opendev system-config
roles [1] are not likely to be accepted [2], this commit adds them
to the sf-infra repository, including the required changes to
support CentOS 7 and 8.

A future change will migrate the existing AFS mirrors to use the
new roles.

[1] - https://review.opendev.org/#/q/status:open+project:opendev/system-config+branch:master+topic:mirror-centos
[2] - http://lists.openstack.org/pipermail/openstack-discuss/2020-July/016022.html

Change-Id: Id6e1e186fe9610da7bb5fe6b79b042b845480beb