mysqld-exporter: use its own SSL certificates
Previously, the mysqld-exporter role used the same SSL certificates
we had for the mariadb-replicated role. This can cause an issue due
to conflicting SELinux labels, which would prevent mariadb from
restarting after the container is created.

Also, add the mysqld_exporter port to the firewalld configuration
in dlrn-db.rdoproject.org.

Change-Id: I32562da00b4863ab795c4dab482754cf2f6612aa