Instance: use Optional Server to indicate if a server is created
This change improves the Instance semantic by replacing the `skip_os_server_task`
boolean with an Optional Server attribute. So that instead of setting an empty
server with a fake image along with a skip_os_server_task to True, an Instance
can simply define a `None Server` to indicate the server task shall be skipped.

Change-Id: Iadce6f8a7bc214708d24ef07808f6e1b20addb3e