Add compression to logrotate config on sf-infra-base
In our systems, it is in general a good idea to enable compression
of rotated logs. While we enable it for several services, the
system-wide configuration has it disabled by default, which may
cause issues for VMs with large /var/log/messages files.

This commit adds a default "compress" line to /etc/logrotate.conf
on the sf-infra-base role.

Change-Id: I9a0517b6b29224ea3e4c6a477f8499daa71576cb