Use OSPO postfix role
The included role is very basic and only needed for the bridge. It
conflicts with an upcoming PR to port mailing-lists from the old repo
and since a renaming is not possible this PR reimplement the current
behavior with the OSPO role.

Since myorigin is set then /etc/mailname is not needed. The rest of the
configuration should be identical or without consequence.

Change-Id: I49f4aa210a62bed7fd0e85820ee8dc830933e7ff