Change order of nodepool builders in sf.io arch file
Currently, sf-config only supports a single nodepool builder in its
gateway.common configuration, so it seems to use the last builder
configured in the arch.yml file. While this is fixed, let's invert
the order to use the main builder host in Vexxhost, instead of the
one in PSI public.

Change-Id: Ib7fa7fe59c2028b48ab3016825cc9b541129fbf2