Bridge: allocate current IP to bridge
Previously, the create-bridge.yaml playbook would use "auto_ip: yes"
to automatically allocate a floating IP to the bridge. This could
cause issues if we rebuild the bridge, since a different IP would
be assigned.

This commit ensures we always use the same IP for the bridge.

Also, change the sf-infra-create-bridge job to run on an existing
container label.

Change-Id: I1ea3d0aa0388fd5362c0295dc768bb10f14bf725