Revert "Pin some OSAS roles to pre-ansible 3.0.0 version"
This reverts commit 02a6e60859aa4fab0f173c9fa0ee886be3bb9853.

Ansible 3.0.0 pulled a version of ansible-base which removed the routing
for the firewalld module. I switched over to the collection name to fix
deployment and tests but unfortunately broke things in communities still
using Ansible 2.9. It was in fact a bug in ansible-base introduced when
the module was moved in the community repo and not an intentional move
towards collections.

Since ansible-base 2.7 is out this bug is fixed. I have reverted all the
changes in OSPO roles and our tests now skip the broken versions. OSPO
roles with tests are now testing Ansible 2.9 & 3.0 to avoid future
breakage. It is now safe to remove this pinning.

Change-Id: I166e0b1d8f5c050fc7a22b27f2f2c1b110f92fc1