Use diffent thresholds for DLRN centos7 and 8 alerts
Currently, monitoring for all DLRN workers was configured to trigger
an alarm if there were no changes for more than 3 days. This is ok
for older branches, which are supported on CentOS 7. However, we want
to be notified earlier for CentOS8-supported branches.

This commit splits the alarm into two: one for CentOS 7 workers (3 days)
and one for CentOS 8 workers (1 day).

Change-Id: Ic993bb9cdad3294dc9592be5a147dce6252aa64d