Update prometheus and prometheus exporter container images
Updated images:
- node_exporter: v1.0.1 to v1.1.2
- blackbox-exporter: v0.18.0 to v0.19.0
- grafana: changed repo from app-sre to dpawlik; bump to v7.5.7
- prometheus: v2.24.1 to v2.27.1
- statsd-expoter: v0.19.0 to v0.20.2
- mysqld-exporter: changed repo from softwarefactory to official; bump
  to 0.13.0-rc0

The app-sre repository for garana is not accessible on quay.io and the
docker.io hub require login to get the newest images.
Temporary created grafana respository, that mirror image v7.5.7 from
docker.io/grafana/grafana and push it to the quay.io/software-factory/grafana.
In the future, container update job will do that.

Change-Id: I94612c5e10a3c229cb144e6e5267a301ff8fe228