Add SECRET_KEY param to Quay configuration
The SECRET_KEY needs to be added into Quay config before some token
is generated. If it is missing after creating a token and
enable mirroring in repository, Quay will raise an error:

    File "/quay-registry/endpoints/api/__init__.py", line 295, in wrapped
      return func(self, namespace, repository, *args, **kwargs)
    File "/quay-registry/endpoints/api/__init__.py", line 466, in wrapped
      resp = func(self, *args, **kwargs)
    File "/quay-registry/endpoints/api/mirror.py", line 237, in get
      logger.warning(
  NameError: name 'logger' is not defined

In that case, you need to drop records in `repomirrorconfig`
and `repomirrorrule` tables related to the repository.

Change-Id: I111376b503eb8c9b61307437e8c4ca780c10bd13