Add floating ip to elk.softwarefactory-project.io
The elk.sf.io will send data to the opensearch.rdoproject.org host and
this host is in different project and it does not use shared local
network, so all the data will be send via public address.
To achieve that, we need to open some port for that host on
opensearch host and without attached floating ip to the host it will be
not possible.

Change-Id: I3494bcc2b4ae221067b87bf72aea000bc22bcdc1