[zuul scrapper] Scrapp when multiple job_name or tenant provided
With this commit, Zuul Log Scrapper will be able to iterate on
multiple Zuul API urls or iterate on multiple job-name.
It means that it is not required to start multiple scrapper
services to pull and push logs from multiple Zuul tenants to the same
endpoint.
Same for --job-name - when multiple job-name provided, it is not
required to run many times same script with only changed param --job-name.

Change-Id: I62946e68bbd41aeee8022267f1643f62f91110d0