Switch RDO Trunk builders to zed tags
This is part of the Yoga branching. At this poing we can diverge
dependencies between zed and yoga, so in this review I'm switching
master builders to zed tags and moving yoga to yoga instead of yoga-uc.

Additionally I'm removing centos9-master-preview builder which is not
used since long time ago.

Note that this needs to be coordinated with [1].

[1] https://review.rdoproject.org/r/c/rdoinfo/+/41455

Change-Id: I635744f704a15b9dbd3871ddda8b9ef293e0a959