create-tripleo-standalone: Move cloud config part in a separate play
The config part for clouds should be in a separate file and the
ansible play should be directly call with a command tasks for it's not
possible to use add_host from zuul-executor for we don't know
baremetal ssh keys during the deployment.

The configure-private-clouds.yaml is used to setup cloud secgroup and
flavors. It could be done only from baremetal hosts for is the only
place where an admin account is defined.

Change-Id: If64ef41098cb2efbd72187332885ef1288d40bbe