[kubernetes] Add loadbalancer MetalLB for ingress; improve Kubernetes role
This commit brings the ingress with MetalLB.
It means, that the ingress host binding is not set to 127.0.0.1
(localhost) and to reach the service deployment, all traffic goes via
loadbalancer or by communicating directly to the ingress http (30080) and
https (30443) ports. The loadbalancer is deployed locally, not from the
cloud provider. More info [1][2].
Also that change improve the Kubernetes ansible role.

[1] https://kubernetes.github.io/ingress-nginx/deploy/baremetal/#a-pure-software-solution-metallb
[2] https://www.cloud-native-everything.com/ingress-and-metallb/

Relates-To: https://softwarefactory-project.io/r/c/software-factory/sf-operator/+/26639

Change-Id: Ia606103dfe6250aac7bd3ee9eed8381b32f78173