[kubernetes] Change ingress provider; bind ingress to hostNetwork
This commit change the ingress provider to kubernetes/ingress-nginx.
The Kind ingress rules does not work correctly on the baremetal
deployment with extra/kubernetes ansible role. It requires to set the
label:

    kubectl label nodes kubernetes.dev ingress-ready=true

for the node, but after apply, ingress controller pod was in
ContainerCreating state, where in events was an error:

    failed to execute iptables-restore: exit status 4

In that case, we modify the baremetal ingress-nginx controller, by
adding:

    - hostPort: 80
    - hostPort: 443
    - hostNetwork: true

To the ingress deployment. After deploying the service, you can reach
them by going: <host ip>:80 or <host_ip>:443.

Change-Id: I1e0d3f3904d1e7af6518b595976b70613a565b5e