Add openstack-k8s-operators project; add more info
The openstack-k8s-operators has been already created, but
it does not configure the pruner.
This commit provides a token for the pruner and configures
pruner to run in that organization.
Also included more details how to add the add new organization
into the Quay and how to generate the application token.

Change-Id: I6ac854fee49087d5dc3747710c0c75028805ab9b