[afs] Port configs from upstream; use httpd-scl; disable apache exporter
The Pip tool can not download packages from our mirror, due
the url is too long. To make it happen, the httpd package needs
to be in version at least 2.4.11, which is not provided by Centos
community, so it requires to use software collection.
This change is disabling apache-exporter role on mirror host, because
it is trying to overwrite packages provided by
opendev-system-config-mirror role.
This commit ports patches from upstream [1][2][3][4]

[1] https://review.opendev.org/c/opendev/system-config/+/850677
[2] https://review.opendev.org/c/opendev/system-config/+/850688
[3] https://review.opendev.org/c/opendev/system-config/+/826974
[4] https://review.opendev.org/c/opendev/system-config/+/758469

Change-Id: I092197a2e890f742fed1b6c9ded831ca7bcc8ec6