[crc] Set CRC version
After creating the CRC server snapshot, the service images related to
the CRC might be updated when the CRC is set to latest, so the
pull-secret.txt file needs to stay on the VM even after starting the
server.
This commit would avoid such situation, because we will set the version
for CRC deployment which would be: <last stable release> - 1.
The error like:

    warning: Container container-00 is unable to start due to an error:
    Back-off pulling image
    "quay.io/openshift-release-dev/ocp-v4.0-art-dev@sha256:afa(...)1bf"

might appear when the service container would be updated.

Change-Id: Ib31981438da3009625d74432083bce60a1f524b7