Add the systemd_unit_inactive rule
We need to track inactive service as well. One service
(opensearch_dashboard)  was inactive for
quite "long" and prevented the backup to run as the backup playbook was
unable to contact the service api.

See for the proof

https://prometheus.monitoring.softwarefactory-project.io/prometheus/graph?g0.expr=node_systemd_unit_state%7Bname%3D%22opensearch-dashboards.service%22%2C%20instance%3D%22managesf.softwarefactory-project.io%3A9100%22%7D&g0.tab=0&g0.stacked=0&g0.show_exemplars=0&g0.range_input=2d&g0.end_input=2023-04-12%2012%3A17%3A08&g0.moment_input=2023-04-12%2012%3A17%3A08

Here we add a new rule for inactive state.

RHOSZUUL-1253

Also move to label zuul-worker-dhall-f37

Change-Id: Ifbca34bc87581440eb2f425b7fa8025d3be2f8a0