Set log_size_max size for podman logs
Sometimes the logs can take over 20GB in few weeks.
By setting 1GB as log_size_max should avoid situation that we are
out of the disk few times per week.
The feature has been added into the podman containers.conf file in
podman 2.2.0 release [1], but on Centos 7, version is below 2.2.0.
According to the libpod.conf man [2], that option should be also
available in podman 1.6.4, but it is located in libpod.conf file.
More info [3].

[1] https://github.com/containers/podman/releases/tag/v2.2.0
[2] https://manpages.debian.org/unstable/podman/libpod.conf.5.en.html
[3] https://github.com/unifi-utilities/unifios-utilities/issues/100

Change-Id: Ic6d01e11606c9526d1880583876d76c4415250ac