Change cloud provider Grafana datasource; mount Grafana DB file
The cloud provider datasource was not working as expected after the
Ansible deployment - we always need to set the authentication
credentials manually. With this commit, we have exported the Grafana
datasources:

    <gen API key in Grafana -> Configuration -> API Keys
    curl -H "Authorization: Bearer KEY" https://host/grafana/api/datasources | jq '' > datasources
    yq -p json -o yaml datasources

and we updated the fields, that has been changed since last version.

It will require to set the basicAuthPassword via Web site (first time) for
datasources, where the basicAuth is enabled, but we also mount the Grafana
library dir, so it would be not required to write the password each
time when the container is restarted.

Change-Id: I05d93e55483759743cff68047042d0f3f49b97f6