Add Opendev Graphite datasource; add AFS dashboard
We would like to have the AFS dashboard inside our Grafana.
All dashboards [1] and datasource from Opendev are available public [2],
so we will just take the data from theirs datasource.
About the dashboards: they are available in Opendev and written in yaml,
due the infra maintaners choose grafyaml tool [3].
Basicly, we upload the dashboard yaml with added "datasource:
Graphite-opendev" key (because we have many datasources), for example:

    (...)
    - title: Partitions
      showTitle: true
      height: 200px
      panels:
        - title: "afs01.dfw /vicepa used"
          datasource: Graphite-opendev
          span: 3
          sparkline:
            full: true
            show: false
          thresholds: "0,90,95"
          format: percent
          targets:
            - target: "stats.gauges.afs.afs01_dfw_openstack_org.part.vicepa.total"
              refId: "A"
              hide: true

remove panels not related to our project and then upload via grafyaml tool:

    grafana-dashboard --grafana-url https://monitoring/grafana/ \
      --grafana-apikey $GRAFANA_APIKEY \
      update \
      afs.yaml

After that, dashboard is available, so we export it via Grafana website
to json file.

[1] https://opendev.org/openstack/project-config/src/branch/master/grafana
[2] https://opendev.org/openstack/project-config/src/branch/master/grafana/datasource.yaml
[3] https://github.com/deliveryhero/grafyaml

Change-Id: I67c2d7d0f2e1e1aab33e2ffe441eb20bd411543c