Add playbook and role for synchronizing image between clouds
It will be needed that same image will be used in few cloud providers.
Also it can happen that same image will be used in few projects in same
cloud provider.
That commit is taking newest image from one cloud provier, check if same
image is present in remote one and if not, will push it to there.
On the end, if there is a need to share image between projects, it will
handle that.

Change-Id: Ie4bac3f4e988b49675906c360cf33c35f6e047c4