Add support for multiple cloud on same host for creating scripts
We would like to run on same host scripts that will handle update
the images on upstream and downstream, where some images will be
extracted, some not. Some other will be shared, some not.
That commit is making a first step to handle that.
This commit also brings few improvements:

- set modify_kubelet_systemd to false because the OKO/next-gen team does not
  see any improvements after the change
- the crc-make-snapshot playbook will create nested and extracted
  version at same time, so we don't need to have many cron jobs

Change-Id: Ib0d1c359318492254f96512561e81ce4b4dd45ef