Add playbook to promote CRC image
We would like to run a tests after creating a new CRC image
(nested/extracted). This playbook will check the image dates
and if it will find that there is a new image available, it will
run tests on that image. When it pass, it will change the image name
to latest one.

Change-Id: Ic04efbbe7f013fbab2bd699e25bdb47fa302115b