Rename CRC image when synchronized and not promoted
On executing promotion process, images with dates will be changed
after to name with '-latest' suffix, when it pass the tests.
In some clouds, we are not doing promotion process for extracted images,
because those images are already promoted in upstream cloud. It means,
that by pulling this image and push to other cloud, name of that image
will also contain '-latest' suffix and each time when the
synchronization will be done, new image with such suffix will be
created.
It will raise an issue, because we will not know which image is really
latest one. In that case, we need to change image name to name with
datetime as we do in promotion process, so there would be just
one image with '-latest' suffix.

Change-Id: I137afd26a0ee6fead99ee06abbf4b3563755985c