Add feature to extracted CRC images that will add ssh keys from magic IP
The extracted CRC image is based on CoreOS, which does not have
cloud-init or ignition installed. It makes a problem to inject
additional ssh pub keys if available.
This commit is taking available pub keys from 169.254.169.254 address
and add it to the authorized_keys via dedicated service, that is
starting right after boot and before kubelet.
Also added support for config-drive, when it is set during instance
creation. The basic user_data that would be parsed looks like:

    #cloud-config
    password: core
    ssh_authorized_keys:
      - ssh-rsa AAAAB(...)Jh0CpE+1xQ4ow==
      - ssh-ed25519 AAAAC3(...)3dsAaY9
      - ecdsa-sha2-nistp256 AAAAE2VjZ(...)uq5QC51C5+WWk=
      - ssh-dss AAAyiM6qiF(...)iG8Wp4O73VUu2bg== someuser@host

Change-Id: I24ec2dd85f6630406e216ebfe60a71488dd36cbd