Ensure that latest image exists on remote cloud; fix debug message
On fresh deployed cloud, the CRC image is not available there, so
the tasks that will rename the "latest" images to name with a timestamp
will fail.
This commit provides condition to verify that remote cloud contains an
image with "latest" tag. If not, it will not rename the image.
Also changed debug message which might be confusing.

Change-Id: Ic67d9b207a12f61349882ad5f076882b48cb0f5a