Update Quay and related services
The currently running Quay service is EOL and it requires to make
upgrade. There is no need to upgrade only for one release - we can do
upgrade directly to the newest one.

Before performing upgrade, we should make a backup:

    mkdir -p /var/data/backup
    podman exec -it postgresql pg_dump -U quay -d quay -c > /var/data/backup/dump-quay
    podman exec -it postgresql pg_dumpall -c -U quay > /var/data/backup/dump-all

Then optionally, we can do also backup of the PostgreSQL directory:

    systemctl stop quay
    systemctl stop quay-mirror
    systemctl stop clair
    systemctl stop postgresql

    tar caf /var/data/backup/backup-dir.tar.gz /var/lib/pgsql/data/

When backup is done, that commit can be merged and the new started
Quay container will do DB migration.

Change-Id: I13c973cbf6878c7d38d98097c09e28ec88b73f96