Remove k1s02 host; fix generating /etc/hosts when server is removed
The k1s02 is based on the Centos 7 which will be soon EOL.
Jobs that are only working on old system should be deprecated. Our other
k1s hosts are based on Fedora system, where one of the k1s host is able to
run the legacy jobs based on Centos 8 and Centos 7 systems until
they are not EOL.
Also need to apply patch proposed in PS [1], because they cannot be
merged separate.

[1] https://softwarefactory-project.io/r/c/software-factory/sf-infra/+/30569

Change-Id: I50fdf1140b63313e15a7e4f5d439b9e761ebac5d