Install ansible-core from the pip repository to the venv; fix scripts
The Centos 9 Stream and RHEL 9 have only Ansible 2.14 available in the
repo. By installing epel-release and epel-release-next, you would be
able to install Ansible in version 7.0, which can not be installed on
RHEL 9.
The next gen role require to install newer openstacksdk than 0.62.0 that
is last working version with Ansible 2.14. It means, that if the host is
running on RHEL 9, currently there is no way to install newer version as
2.14. That breaks the CRC scripts that are triggering Ansible to create
new CRC images.

This commit is installing ansible-core package from the pip repository
to the venv instead of relying of rpm packages.
Also improve scripts to be more similar to each other.

Change-Id: If997a6cd21058db986b12077f5084955898b0e9e