create-hosts.yaml: change how host are created
We have an issue with f39 and ansiblesdk 2.1.0 when we use
cloud.openstack.server module. The issue is ansible timeout with the
following error:

"msg": "Timeout waiting for the floating IP to be attached."

the issue is the servers are up and running, ansible can reach them.
The root cause seems to be the usage of auto_ip and floating_ips [1]
with wait option, these options should be removed in a following patch for:

'Options auto_ip, floating_ip_pools and floating_ips interact in non-obvious ways and undocumentable depth. For explicit and safe attaching and detaching of floating ip addresses use module openstack.cloud.resource instead.'

This change ensure cloud.openstack.server_info is used to collect all
instances in the tenant and install the host only if the server is not available

This is also quicker than executing cloud.openstack.server for each host

a run d/s was 20 mins 33 secs and it's around 1 min with this change

Finally, add loop_control label [2] to reduce ansible output

[1] https://docs.ansible.com/ansible/latest/collections/openstack/cloud/server_module.html#parameter-auto_ip
[2] https://docs.ansible.com/ansible/latest/playbook_guide/playbooks_loops.html#limiting-loop-output-with-label

Change-Id: I016356811623de73a18002952088a1fc8ff14a03