generate-etc-hosts: change the way we collect instances ips
Due to issue with auto_ip with f39 and openstacksdk 2.1.0, we will
change the way servers are created for we will only create server if
they are not in active state (ansible don't recreate server if it
exists so it won't change anything on host creation)

It won't be possible to generate /etc/hosts based on
openstack.cloud.server result anymore.

This change add a task in playbooks/tasks/create-hosts.yaml to collect
publics ip for the hosts on infra tenants, and collect private ips for
sf instances (nodepool-builder, ze0*, zm0* and zs). The /etc/hosts is
then populated with this ip if the role generate-etc-hosts is
included (today only brigde and prometheus instances)

The variables to collect tenants are added in host_vars/bridge file
and are provided to the role with managed_instances variable to
facilitate usage d/s

Change-Id: I5080426ce326eafd29b908d8b6523b2a64859ddf