generate-etc-hosts: move from system to bridge directory
This change move system/generate-etc-hosts in
bridge/generate-etc-hosts to fix the issue where all
sf-infra-configure-hosts-* jobs are run for we have a file matcher on
'system/*' roles. This change will speedup the process when we add a
new instance to run only site_bridge and the site file where the host
was added.

Change-Id: I400aeafd521fa3a329cbd8a2fa01f69146b7db81