create-hosts.yaml: change how fip are attached to server
With cloud.openstack.server module is it not recommanded to use
anymore auto_ip and floating_ip parameters for they have umpredictable
behaviors.

This change set auto_ip to False (default is True) even if it should
not be used and add floating_ip boolean to defined if a server should
have a floating ip.

- modify tasks/create-hosts.yaml playbook to handle server creation
and deletion for servers
- create tasks/manage-server.yaml to manage fip and server creations
and deletions
- modify Infra/Server to match this change

Change-Id: Ie3761cb3f2c5028ef640d73dabef98a4e446a7f7