observability: add fluent bit deployment role
Deploy a fluent bit forwarder in the same namespace as a Software
Factory instance. Fluent bit will forward the logs to the designated
Loki instance.

This is meant for Software Factories deployed on clusters where it
is impossible to forward logs by installing Promtail on the nodes
(as a service or as a daemonset).

Change-Id: I7a2ff798ab9474e27d7e478c86156ae84a63b225