Observability: add role to deploy prometheus + thanos sidecar
This role assumes the kube context is allowed to deploy prometheus CRs.
The role can handle the creation and configuration of a serviceaccount
that can be used with the prometheus instance.
If an object storage configuration is provided, the instance will be
deployed with a thanos sidecar to forward metrics data to object
storage.
Extra scrape configs can be provided.

Change-Id: I1773c46b83880cc4f5e8830240388920e3e9a72b