Enable httpd24 environment to run htcacheclean command
The crontab entry was not working properly after moving to the httpd24
package, because the htcacheclean binary was not available in the path
that is provided.
With current way, we need to enable httpd24 environment, then run the
command (which is also available: /opt/rh/httpd24/root/usr/sbin/htcacheclean).

Change-Id: Id9a0d160d27b68c77ca4723c12e5a62d6fbbed2c