Take created_at field instead of updated_at for cleaning DOWN ports
The next gen cleanup ports script was taking date from a field
"updated_at", which is wrong, due even small operation done on the
port, trunk network etc. is changing the date field, that later it is
skipped by the script.
This commit is changing to take value from a field: `created_at`,
because we should base on that.

Change-Id: If6c0c2286e90a2facedab8fdf3af8df411b3adec