Move CI test job to use Centos 9 Stream instead of Centos 8 Stream
We are upgrading the hosts to newer system release.
Soon the C8S would be EOL, so let's move to use Centos 9 Stream.
Also commit contains fix for Grafana to make the test pass and
remove task that is installing: centos-release-openstack-victoria,
because python3-openstacksdk package is available in Fedora repository.

Change-Id: I5dd940d1871b90ac6e3a0304193cd9e4b64bf749