Set dedidated network for Quay
Currently we spotted an error in Quay, when we want to reach the
service from one cloud provider. The error is:

    error: build error: Failed to push image: trying to reuse blob sha256:aa(...)
    at destination: pinging container registry quay.rdoproject.org: Get
    "https://quay.rdoproject.org/v2/": net/http: TLS handshake timeout

The Quay service can not be running with parameter --network host, due
the port binding in the service is 8080 or 8443, which will require
to run socat or http proxy to redirect to the service. It raises some
issues: with one solution web site is working, but the container images
can not be downloaded, another got opposite error.

That commit is making a dedicated network for Quay service, that sets
MTU to 1300 as default (minimum value for the VM in cloud provider is
1500).

Change-Id: I79784efef6b72a09f73186f9c89a76aae301c9e2