Configure logscraper02.openstack.org; drop logscraper01.openstack.org
Our team is maintaining the logscraper01.openstack.org host,
which is has been spawned by the OpenDev team, but because the
host role is to compute and send the logs to the OpenDev Opensearch,
where we are also maintaning it, we also take the responsibility for
that host.
The logscraper01 is based on Centos 8 Stream which is EOL soon.
The new host (logscraper02.openstack.org) is now running on Centos 9 Stream.

More info about ci-log-processing project [1].

[1] https://docs.openstack.org/infra/ci-log-processing/

Change-Id: I55d62bf6565b352782d73f865baca5ca2ab800c4