Add trunkbackup user to backup server
In order to implement a backup solution for the new RDO Trunk service I
am requesting to create a user trunkbackup and a directory owned by it
in the backup volume. I will use that directory to push backups
periodically from trunk-builder-centos9 and dlrn-db-centos9 servers.

In order to use rsync over ssh I'm also requesting to inject an
authorized ssh key for the new user.

Change-Id: If40b7b3e6692164e1afb583c80c0d21a3693bd10