Always try to pull exporters new image when image is latest
Some images we have defined in group_vars, some others we overwrite in
host_vars. The current issue is, that the container_images hashes are
not merged. It means, that on Quay or OpenSearch RDO VMs, node-exporter
service that is running there is using version: latest. Therefore,
the service will be outdated comparing to what is set globaly.

Change-Id: Ic3046ee76aa1ca1b5a4776e3140e39c07cd61440