Change Prometheus datasource scrape interval to 1m
This commit is fixing an issue in node exporter full dashboard, where
some visualizations are not displayed and it shows "No data".
The issue is not in the Prometheus, but in the Grafana. The default
options in the Prometheus are:

    # /etc/prometheus/prometheus.yaml
    (...)
    global:
      evaluation_interval: "1m"
      scrape_interval: "1m"
      scrape_timeout: "10s"

The same option (scrape interval) should be set also in the Grafana datasource.

Change-Id: I19887adb0beb6d6d5d68b239c56658f3bd76ec85